refactor to redirect to auction creation
