+++ auction-ui/src/routes/my-products/MyProducts.jsx
@@ -1,6 +1,7 @@
 import { useState, useEffect } from "react";
 import ProductCard from "./components/ProductCard";
-import AddProduct from "./components/AddProduct";
+import AddProduct from "../new-product/components/AddProduct";
+import { Link } from "react-router-dom";
 import axios from "axios";
 
 export default function MyProducts({currentUser}) {
@@ -23,13 +24,18 @@ export default function MyProducts({currentUser}) {
             <div className="row">
                 <div className="col-8">
                     <h4 className="pt-4">List of saved products to auction</h4>
+                    
                     {products && products.map(product => {
                         return <ProductCard product={product} key={product.id}/>
                     })}
                 </div>
-                <div className="col">
-                    <h4 className="pt-4 ">Add a new product</h4>
-                    <AddProduct currentUser={currentUser} setProducts={setProducts}/>
+                <div className="col-4 mt-5">
+                    <Link 
+                        className="btn btn-dark" 
+                        to="/new-product"
+                    >
+                        Add a new product
+                    </Link>
                 </div>
             </div>
         </div>
